[REFACTOR] Constant 리팩토링
- 글로벌하게 관리하도록 수정했습니다.
