+++ src/main/java/plub/plubserver/common/constant/GlobalConstants.java
@@ -1,9 +1,17 @@
-package plub.plubserver.domain.report.config;
+package plub.plubserver.common.constant;
 
-public class ReportConstant {
+public class GlobalConstants {
+
+    /**
+     * Account
+     */
+    public static final int NICKNAME_CHANGE_LIMIT = 3;
+
+    /**
+     * Report
+     */
     public static final int REPORT_WARNING_PUSH_COUNT = 6;
     public static final int REPORT_PLUBBING_PAUSE_COUNT = 18;
-
     public static final int REPORT_ACCOUNT_WARNING_PUSH_COUNT = 1;
     public static final int REPORT_ACCOUNT_PAUSED_COUNT = 3;
     public static final int REPORT_ACCOUNT_BAN_COUNT = 6;
